NEW: Download Attachment from DB FIX: JPA mapping Email-Attachment

+++ src/main/java/com/engagepoint/university/messaging/dao/generic/GenericDAO.java
@@ -6,7 +6,8 @@ import java.util.List;
 
 public interface GenericDAO<DTOType extends BaseDTO> {
 
-    public DTOType getById(Integer id);
+    public DTOType getById(Long id);
+
     public List<DTOType> getAll();
     public void save(DTOType DTOType);
     public void update(DTOType DTOType);
