+++ src/main/java/com/engagepoint/university/messaging/dao/specific/impl/SmsDAOImpl.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 public class SmsDAOImpl implements SmsDAO {
     @Override
-    public SmsDTO getById(Integer id) {
+    public SmsDTO getById(Long id) {
         EntityManagerUtil.getEntityManager().getTransaction().begin();
         Sms sms = EntityManagerUtil.getEntityManager().find(Sms.class, id);
         SmsDTO smsDTO = new Converter().convert(sms);
@@ -95,7 +95,7 @@ public class SmsDAOImpl implements SmsDAO {
     @Override
     public void deleteIdList(List<Long> idList) {
         EntityManagerUtil.getEntityManager().getTransaction().begin();
-        EntityManagerUtil.getEntityManager().createNamedQuery(Sms.DELETE_SMS_LIST).setParameter(Sms.PARAM_IDS_LIST,idList).executeUpdate();
+        EntityManagerUtil.getEntityManager().createNamedQuery(Sms.DELETE_SMS_LIST).setParameter(Sms.PARAM_IDS_LIST, idList).executeUpdate();
         EntityManagerUtil.getEntityManager().getTransaction().commit();
     }
 }
